requirements.txt
Django==4.0
psycopg2_binary==2.9.2
djangorestframework-3.15.2
django-filter-24.3